Separate replay debugger adapter from extension
