+++ packages/salesforcedx-apex-replay-debugger/README.md
@@ -1,0 +1,8 @@
+# salesforcedx-apex-replay-debugger
+This repository contains a debug adapter that implements VS Code Debug Protocol.
+It is a standalone executable that translates between VS Code and Salesforce
+Apex Debugger.
+
+## Resources
+* [VS Code Debug Protocol in a Nutshell](https://code.visualstudio.com/docs/extensionAPI/api-debugging)
+* [VS Code Debug Protocol Reference](https://github.com/Microsoft/vscode-debugadapter-node/blob/master/protocol/src/debugProtocol.ts)
