+++ packages/salesforcedx-apex-replay-debugger/src/core/logContextUtil.ts
