+++ packages/salesforcedx-apex-replay-debugger/src/states/debugLogState.ts
