+++ packages/salesforcedx-apex-replay-debugger/src/states/frameEntryState.ts
