+++ packages/salesforcedx-apex-replay-debugger/src/states/frameState.ts
