+++ packages/salesforcedx-apex-replay-debugger/src/states/frameStateUtil.ts
