+++ packages/salesforcedx-apex-replay-debugger/src/states/logEntryState.ts
