+++ packages/salesforcedx-apex-replay-debugger/src/states/variableAssignmentState.ts
