+++ packages/salesforcedx-apex-replay-debugger/src/states/variableBeginState.ts
