+++ packages/salesforcedx-apex-replay-debugger/test/unit/states/frameEntryState.test.ts
