+++ packages/salesforcedx-apex-replay-debugger/test/unit/states/logEntryState.test.ts
