+++ packages/salesforcedx-apex-replay-debugger/test/unit/states/variableAssignmentState.test.ts
