+++ packages/salesforcedx-apex-replay-debugger/test/unit/states/variableBeginState.test.ts
