+++ packages/salesforcedx-vscode-apex-replay-debugger/package.json
@@ -22,12 +22,9 @@
   },
   "categories": ["Debuggers"],
   "dependencies": {
-    "@salesforce/salesforcedx-utils-vscode": "42.17.0",
+    "@salesforce/salesforcedx-apex-replay-debugger": "42.17.0",
     "path-exists": "3.0.0",
-    "request-light": "0.2.1",
-    "vscode-debugadapter": "1.28.0",
-    "vscode-debugprotocol": "1.28.0",
-    "vscode-uri": "1.0.1"
+    "request-light": "0.2.1"
   },
   "devDependencies": {
     "@types/chai": "^4.0.0",
@@ -36,15 +33,9 @@
     "@types/path-exists": "^1.0.29",
     "@types/sinon": "^2.3.2",
     "chai": "^4.0.2",
-    "cross-env": "5.0.4",
     "mocha": "3.2.0",
-    "mocha-junit-reporter": "^1.13.0",
-    "mocha-multi-reporters": "^1.1.4",
-    "nyc": "^11.0.2",
     "sinon": "^2.3.6",
-    "typescript": "2.6.2",
-    "vscode": "1.1.17",
-    "vscode-debugadapter-testsupport": "1.28.0"
+    "vscode": "1.1.17"
   },
   "extensionDependencies": [
     "salesforce.salesforcedx-vscode-apex",
@@ -59,10 +50,8 @@
     "clean":
       "shx rm -rf node_modules && shx rm -rf out && shx rm -rf coverage && shx rm -rf .nyc_output",
     "postinstall": "node ./node_modules/vscode/bin/install",
-    "test":
-      "./node_modules/.bin/cross-env VSCODE_NLS_CONFIG={} ./node_modules/.bin/nyc ./node_modules/.bin/_mocha --recursive out/adaptertest --reporter mocha-multi-reporters && node ./node_modules/vscode/bin/test",
-    "test:unit":
-      "./node_modules/.bin/cross-env VSCODE_NLS_CONFIG={} ./node_modules/.bin/nyc ./node_modules/.bin/_mocha --recursive out/adaptertest/unit --reporter mocha-multi-reporters"
+    "test": "node ./node_modules/vscode/bin/test",
+    "test:unit": "node ./node_modules/vscode/bin/test"
   },
   "activationEvents": [
     "onDebug",
@@ -176,7 +165,8 @@
       {
         "type": "apex-replay",
         "label": "Apex Replay Debugger",
-        "program": "./out/src/adapter/apexReplayDebug.js",
+        "program":
+          "./node_modules/@salesforce/salesforcedx-apex-replay-debugger/out/src/adapter/apexReplayDebug.js",
         "runtime": "node",
         "languages": ["apex"],
         "variables": {
@@ -231,8 +221,5 @@
         }
       }
     }
-  },
-  "nyc": {
-    "reporter": ["text-summary", "lcov"]
   }
 }
