+++ packages/salesforcedx-vscode-apex-replay-debugger/src/commands/apexExecutionOverlayActionCommand.ts
@@ -1,17 +1,18 @@
 /*
- * Copyright (c) 2017, salesforce.com, inc.
+ * Copyright (c) 2018, salesforce.com, inc.
  * All rights reserved.
  * Licensed under the BSD 3-Clause license.
  * For full license text, see LICENSE.txt file in the repo root or https://opensource.org/licenses/BSD-3-Clause
  */
 
-import { BaseCommand } from '@salesforce/salesforcedx-utils-vscode/out/src/requestService';
+import { BaseCommand } from '@salesforce/salesforcedx-apex-replay-debugger/node_modules/@salesforce/salesforcedx-utils-vscode/out/src/requestService';
 
 export interface ApexExecutionOverlayFailureResult {
   message: string;
   errorCode: string;
   fields: string[];
 }
+
 export interface ApexExecutionOverlaySuccessResult {
   id: string;
   success: boolean;
@@ -22,7 +23,6 @@ export interface ApexExecutionOverlaySuccessResult {
 export class ApexExecutionOverlayActionCommand extends BaseCommand {
   private readonly apiPath = 'services/data/v43.0/tooling/sobjects';
   private readonly commandName = 'ApexExecutionOverlayAction';
-  // The request string is basically the json string of the argumnets, if there are any
   private readonly requestString: string | undefined;
   private readonly actionObjectId: string | undefined;
 
@@ -46,12 +46,10 @@ export class ApexExecutionOverlayActionCommand extends BaseCommand {
     }
   }
 
-  // For this particular message the query string should be undefined
   public getQueryString(): string | undefined {
     return this.queryString;
   }
 
-  // The requestBody is going to contain the JSON string of all arguments
   public getRequest(): string | undefined {
     return this.requestString;
   }
