+++ packages/salesforcedx-vscode-apex-replay-debugger/src/commands/queryExistingOverlayActionIdsCommand.ts
@@ -1,11 +1,11 @@
 /*
- * Copyright (c) 2017, salesforce.com, inc.
+ * Copyright (c) 2018, salesforce.com, inc.
  * All rights reserved.
  * Licensed under the BSD 3-Clause license.
  * For full license text, see LICENSE.txt file in the repo root or https://opensource.org/licenses/BSD-3-Clause
  */
 
-import { BaseCommand } from '@salesforce/salesforcedx-utils-vscode/out/src/requestService';
+import { BaseCommand } from '@salesforce/salesforcedx-apex-replay-debugger/node_modules/@salesforce/salesforcedx-utils-vscode/out/src/requestService';
 
 export interface QueryOverlayActionIdsSuccessResult {
   size: number;
@@ -37,13 +37,10 @@ export class QueryExistingOverlayActionIdsCommand extends BaseCommand {
     return this.apiPath;
   }
 
-  // For this particular message the query string will be the combination of the queryString passed
-  // into the super with the single quoted userId appeneded at the end
   public getQueryString(): string | undefined {
-    return this.queryString + "'" + this.userId + "'";
+    return `${this.queryString}'${this.userId}'`;
   }
 
-  // For this particular message the requestString should be undefined
   public getRequest(): string | undefined {
     return undefined;
   }
