+++ packages/salesforcedx-vscode-apex-replay-debugger/src/messages/i18n.ts
@@ -19,20 +19,12 @@ export const messages = {
   config_name_text: 'Launch Apex Replay Debugger',
   session_language_server_error_text:
     'Apex language server could not provide information about valid breakpoints.',
-  session_started_text: 'Apex Replay Debugger session started for log file %s.',
-  session_terminated_text: 'Apex Replay Debugger session terminated.',
-  no_log_file_text:
-    'The log file either is missing or does not have any log lines in it.',
-  incorrect_log_levels_text:
-    'The log must be generated with log categories Apex code at the FINEST level and Visualforce at the FINER level.',
   up_to_five_checkpoints:
     'You have %s of the maximum 5 enabled checkpoints. Remove or disable checkpoints until 5 or fewer are active.',
   checkpoints_can_only_be_on_valid_apex_source:
     'Checkpoints can be set on a valid line of Apex source. Invalid checkpoint location: URI=%s, Line=%s',
   local_source_is_out_of_sync_with_the_server:
     "The local source is out of sync with the server. Push any changes you've made locally to your org, and pull any changes you've made in the org into your local project.",
-  checkpoints_successfully_uploaded:
-    'All checkpoints have been successfully uploaded to the server.',
   sfdx_create_checkpoints_start: 'Starting SFDX: Create Checkpoints',
   checkpoint_creation_status_org_info:
     'Step 1 of 6: Retrieving org information',
@@ -50,7 +42,6 @@ export const messages = {
   // These strings are going to be re-worked to become better, Salesforce appropriate, error messages.
   cannot_determine_workspace:
     'Unable to determine workspace folders for workspace',
-  cannot_delete_existing_checkpoint: 'Cannot delete existing checkpoint',
   unable_to_parse_checkpoint_query_result:
     'Unable to parse checkpoint query result',
   unable_to_retrieve_active_user_for_sfdx_project:
@@ -59,8 +50,6 @@ export const messages = {
     'Unable to query for existing checkpoints',
   unable_to_load_vscode_core_extension:
     'unable to load salesforce.salesforcedx-vscode-core extension',
-  unable_to_remove_checkpoint: 'Unable to remove checkpoint',
-  unable_to_create_checkpoint: 'Unable to create checkpoint.',
   no_line_breakpoint_information_for_current_project:
     'There is no line breakpoint informatin for the current project',
   line_breakpoint_information_success:
